requirements.txt
streamlit==1.32.0
google-cloud-vision==3.4.4
openai==1.12.0
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.18.0 